refactor(orchestration): extract TaskApprovalService
Two methods, 176 body lines, moved verbatim. 1570 -> 1399 lines. The plan named
two collaborators and missed two: approveTask calls
questionFlow.assertCoordinatorOwnership three times and, since Task 12,
lifecycle.getTask. Its port list also named logger, which nothing uses.

One receiver rewrite, and it removes a microtask hop rather than adding one.
`await this.getTask(...)` was reaching the facade's own delegation, which has
forwarded to TaskLifecycleService since Task 12; on the service it points
straight at the collaborator. Task 12 added that hop on this exact path and
every fixture stayed green, so the approveTask path does not observe hop depth
-- and removing it leaves all 24 fixtures byte-identical, confirming it.

approveTask opens several mutate blocks in sequence, never nested, with
reserveProposedWorkerSession, ensureReservedWorkerSession and dispatchWorkerTask
deliberately between them. Verified still outside every critical section.

Noted for the whole-branch review, not fixed here: approveTask checks
canStartParallelTask inside a mutate but never calls claimParallelStart, while
both delegation paths do. Whether that asymmetry is a real TOCTOU hole is a
behaviour question, not a move question.
